refactor: contants.js를 constant 폴더로 따로 관리
