+++ src/js/utils/random.js
@@ -1,4 +1,4 @@
-import { GAME_NUMBERS } from './constants.js';
+import { GAME_NUMBERS } from '../constant/index.js';
 
 export const generateRandomNumber = () =>
   Math.floor(Math.random() * GAME_NUMBERS.MAX_RANDOM_NUMBER);
